ScenarioConstellation: correct int rounding issue

+++ Scenario/ScenarioConstellation.py
@@ -156,7 +156,7 @@ class ScenarioConstellation:
         self.define_constellation_orbits()
 
         # Check if satellites volume is known, otherwise an estimate is provided
-        if int(self.sat_volume.value) == 0:
+        if float(self.sat_volume.value) == 0:
             logging.info("Estimating satellite volume...")
             self.estimate_satellite_volume()
